Added more PDOException try/catches

+++ src/Projection/PdoEventStoreProjector.php
@@ -335,12 +335,20 @@ WHERE name = ?
 EOT;
 
         $statement = $this->connection->prepare($sql);
-        $statement->execute([
-            json_encode($this->streamPositions, \JSON_FORCE_OBJECT),
-            json_encode($this->state, \JSON_FORCE_OBJECT),
-            $this->status->getValue(),
-            $this->name,
-        ]);
+        try {
+            $statement->execute([
+                json_encode($this->streamPositions, \JSON_FORCE_OBJECT),
+                json_encode($this->state, \JSON_FORCE_OBJECT),
+                $this->status->getValue(),
+                $this->name,
+            ]);
+        } catch (PDOException $exception) {
+            // ignore and check error code
+        }
+
+        if ($statement->errorCode() !== '00000') {
+            throw RuntimeException::forStatementErrorInfo($statement->errorInfo());
+        }
 
         try {
             $this->eventStore->delete(new StreamName($this->name));
@@ -357,7 +365,15 @@ EOT;
 UPDATE $this->projectionsTable SET status = ? WHERE name = ?;
 EOT;
         $statement = $this->connection->prepare($stopProjectionSql);
-        $statement->execute([ProjectionStatus::IDLE()->getValue(), $this->name]);
+        try {
+            $statement->execute([ProjectionStatus::IDLE()->getValue(), $this->name]);
+        } catch (PDOException $exception) {
+            // ignore and check error code
+        }
+
+        if ($statement->errorCode() !== '00000') {
+            throw RuntimeException::forStatementErrorInfo($statement->errorInfo());
+        }
 
         $this->status = ProjectionStatus::IDLE();
     }
@@ -378,7 +394,15 @@ EOT;
 DELETE FROM $this->projectionsTable WHERE name = ?;
 EOT;
         $statement = $this->connection->prepare($deleteProjectionSql);
-        $statement->execute([$this->name]);
+        try {
+            $statement->execute([$this->name]);
+        } catch (PDOException $exception) {
+            // ignore and check error code
+        }
+
+        if ($statement->errorCode() !== '00000') {
+            throw RuntimeException::forStatementErrorInfo($statement->errorInfo());
+        }
 
         if ($deleteEmittedEvents) {
             try {
@@ -634,7 +658,15 @@ SELECT position, state FROM $this->projectionsTable WHERE name = ? LIMIT 1;
 EOT;
 
         $statement = $this->connection->prepare($sql);
-        $statement->execute([$this->name]);
+        try {
+            $statement->execute([$this->name]);
+        } catch (PDOException $exception) {
+            // ignore and check error code
+        }
+
+        if ($statement->errorCode() !== '00000') {
+            throw RuntimeException::forStatementErrorInfo($statement->errorInfo());
+        }
 
         $result = $statement->fetch(PDO::FETCH_OBJ);
 
@@ -682,7 +714,11 @@ UPDATE $this->projectionsTable SET locked_until = ?, status = ? WHERE name = ? A
 EOT;
 
         $statement = $this->connection->prepare($sql);
-        $statement->execute([$lockUntilString, ProjectionStatus::RUNNING()->getValue(), $this->name, $nowString]);
+        try {
+            $statement->execute([$lockUntilString, ProjectionStatus::RUNNING()->getValue(), $this->name, $nowString]);
+        } catch (PDOException $exception) {
+            // ignore and check error code
+        }
 
         if ($statement->rowCount() !== 1) {
             if ($statement->errorCode() !== '00000') {
@@ -707,7 +743,15 @@ UPDATE $this->projectionsTable SET locked_until = NULL, status = ? WHERE name = 
 EOT;
 
         $statement = $this->connection->prepare($sql);
-        $statement->execute([ProjectionStatus::IDLE()->getValue(), $this->name]);
+        try {
+            $statement->execute([ProjectionStatus::IDLE()->getValue(), $this->name]);
+        } catch (PDOException $exception) {
+            // ignore and check error code
+        }
+
+        if ($statement->errorCode() !== '00000') {
+            throw RuntimeException::forStatementErrorInfo($statement->errorInfo());
+        }
 
         $this->status = ProjectionStatus::IDLE();
     }
@@ -723,12 +767,20 @@ WHERE name = ?
 EOT;
 
         $statement = $this->connection->prepare($sql);
-        $statement->execute([
-            json_encode($this->streamPositions, \JSON_FORCE_OBJECT),
-            json_encode($this->state, \JSON_FORCE_OBJECT),
-            $lockUntilString,
-            $this->name,
-        ]);
+        try {
+            $statement->execute([
+                json_encode($this->streamPositions, \JSON_FORCE_OBJECT),
+                json_encode($this->state, \JSON_FORCE_OBJECT),
+                $lockUntilString,
+                $this->name,
+            ]);
+        } catch (PDOException $exception) {
+            // ignore and check error code
+        }
+
+        if ($statement->errorCode() !== '00000') {
+            throw RuntimeException::forStatementErrorInfo($statement->errorInfo());
+        }
     }
 
     private function prepareStreamPositions(): void
@@ -740,7 +792,15 @@ EOT;
 SELECT real_stream_name FROM $this->eventStreamsTable WHERE real_stream_name NOT LIKE '$%';
 EOT;
             $statement = $this->connection->prepare($sql);
-            $statement->execute();
+            try {
+                $statement->execute();
+            } catch (PDOException $exception) {
+                // ignore and check error code
+            }
+
+            if ($statement->errorCode() !== '00000') {
+                throw RuntimeException::forStatementErrorInfo($statement->errorInfo());
+            }
 
             while ($row = $statement->fetch(PDO::FETCH_OBJ)) {
                 $streamPositions[$row->real_stream_name] = 0;
@@ -769,7 +829,15 @@ EOT;
 SELECT real_stream_name FROM $this->eventStreamsTable $where;
 EOT;
             $statement = $this->connection->prepare($sql);
-            $statement->execute($params);
+            try {
+                $statement->execute($params);
+            } catch (PDOException $exception) {
+                // ignore and check error code
+            }
+
+            if ($statement->errorCode() !== '00000') {
+                throw RuntimeException::forStatementErrorInfo($statement->errorInfo());
+            }
 
             while ($row = $statement->fetch(PDO::FETCH_OBJ)) {
                 $streamPositions[$row->real_stream_name] = 0;
